refactor: extract StatementsListContent into its own component
